Capstone Project/Capstone Project/requirements.txt
selenium==4.18.1
pytest==8.1.1
pytest-html==4.1.1
webdriver-manager==4.0.1
configparser==6.0.0
pandas==2.2.1
undetected-chromedriver>=3.5.5
robotframework>=6.1.1
robotframework-seleniumlibrary>=6.2.0